+++ task-x/src/pages/TaskForm/index.tsx
@@ -1,5 +1,4 @@
 import {
-  CSSProperties,
   ChangeEvent,
   Fragment,
   FunctionComponent,
@@ -11,8 +10,6 @@ import {
   ContainerCard,
   TaskAlertNotification,
   ContainerCardProps,
-  CustomButtonAction,
-  CustomInputText,
 } from "../../components";
 import {
   flexCol,
@@ -24,8 +21,9 @@ import {
 import { ENGLISH } from "../../constant";
 import { Tasks, TypeStatus } from "../../@types/task";
 import { getFormattedDate } from "../../utils";
+import { TaskFormComponent } from "./TaskFormComponent";
 
-const { LABEL, TASK, INPUT, FORM } = ENGLISH;
+const { LABEL, TASK } = ENGLISH;
 
 export interface CreateTaskListProps {
   addTask: (tasks: Tasks) => void;
@@ -48,7 +46,7 @@ export const CreateTaskList: FunctionComponent<CreateTaskListProps> = ({
     status: "completed",
     dueDate: "",
   });
-  const { title, errorTitle, errorDescription, description } = task;
+  const { title, description } = task;
   const selection = useRef<HTMLSelectElement>(null);
 
   const currentDate = getFormattedDate();
@@ -212,17 +210,6 @@ export const CreateTaskList: FunctionComponent<CreateTaskListProps> = ({
     },
   ];
 
-  const disableContinue = !!(
-    errorDescription ||
-    errorTitle ||
-    title.trim() === "" ||
-    description!.trim() === "" ||
-    description!.length < 10 ||
-    title.length < 10 ||
-    title.length > 25 ||
-    (errorDescription && errorDescription.length < 25)
-  );
-
   return (
     <Fragment>
       <div className={flexCol}>
@@ -252,77 +239,19 @@ export const CreateTaskList: FunctionComponent<CreateTaskListProps> = ({
           </div>
         )}
         <div className=" flex flex-1" style={{ borderRadius: 8 }}>
-          <Fragment>
-            <div
-              className="container py-4 shadow-md md:shadow-2xl mt-2"
-              style={{ borderRadius: 8 }}
-            >
-              <div className="p-6">
-                <h1 className="text-xl font-semibold mb-4 text-gray-700">
-                  {FORM.CREATE_TASK_TITLE_LABEL}
-                </h1>
-                <p className="text-gray-600 dark:text-gray-300 mb-6">
-                  {FORM.CREATE_TASK_SUB_LABEL}
-                </p>
-                <Fragment>
-                  <div className="mb-4">
-                    <CustomInputText
-                      onBlur={handleOnBlurTitle}
-                      spaceBetweenText={"p-2"}
-                      value={title}
-                      error={errorTitle}
-                      onFocus={handleOnFocusTitle}
-                      onChange={handleOnChangeTitle}
-                      placeholder={INPUT.TITLE_PLACE_HOLDER}
-                    />
-                  </div>
-                  <div className="mb-4">
-                    <select
-                      className="border p-2 rounded w-full"
-                      ref={selection}
-                    >
-                      <option value={"important"}>
-                        {FORM.OPTION_IMPORTANT_LABEL}
-                      </option>
-                      <option value={"complete"}>
-                        {FORM.OPTION_COMPLETE_LABEL}
-                      </option>
-                    </select>
-                  </div>
-                  <div className="mb-4">
-                    <CustomInputText
-                      onBlur={handleOnBlurDescription}
-                      spaceBetweenText={"p-2"}
-                      value={description}
-                      error={errorDescription}
-                      onFocus={handleOnFocusDescription}
-                      onChange={handleOnChangeDescription}
-                      placeholder={INPUT.DESCRIPTION_PLACE_HOLDER}
-                    />
-                  </div>
-                  <div className="flex justify-center items-center">
-                    <CustomButtonAction
-                      disabledContinue={disableContinue}
-                      textColor={"white"}
-                      iconStyle={{
-                        icon: "add-circle",
-                        size: 20,
-                      }}
-                      label="Add new task"
-                      customStyle={buttonStyle}
-                      onPress={handleAddTask}
-                    />
-                  </div>
-                </Fragment>
-              </div>
-            </div>
-          </Fragment>
+          <TaskFormComponent
+            handleOnBlurTitle={handleOnBlurTitle}
+            handleOnBlurDescription={handleOnBlurDescription}
+            handleOnChangeTitle={handleOnChangeTitle}
+            handleOnChangeDescription={handleOnChangeDescription}
+            handleOnFocusTitle={handleOnFocusTitle}
+            handleOnFocusDescription={handleOnFocusDescription}
+            handleSubmit={handleAddTask}
+            value={task}
+            ref={selection}
+          />
         </div>
       </div>
     </Fragment>
   );
-};
-const buttonStyle: CSSProperties = {
-  backgroundColor: colorGreen[400],
-  borderRadius: 12,
 };
